+++ app/processing/payment/calculate-payment-schedule/index.js
@@ -1,4 +1,4 @@
-const calculatePaymentSchedule = require('./calculate-schedule')
+const calculatePaymentSchedule = require('./calculate-payment-schedule')
 
 module.exports = {
   calculatePaymentSchedule
